Use edit tags instead of edit summaries to detect reverts (#58)
Use edit tags to detect reverts and other types of edits that should be
ignored for contests. This involves a refactoring of the config files to
also include one common default.yml config file, which can be overridden
if needed. This default file includes the tags that we normally don't
want to care about; including tags for revert edits, reverted edits,
undo edits, and various automated Wikidata edits (most notably
QuickStatements).

This fixes #45.