feature:Unit Tests for Site.py (#85)
It confirms that standard prefixes like 'en' and 'de' are properly
registered, while also validating the behavior with edge cases such as
an empty string (''), None, capitalized forms like 'UPPERcase', and
prefixes with leading/trailing whitespace like ' en '. The tests also
verify that the class remains working when unknown or invalid prefixes
(like 'zz', 'not_in_map', or the integer 42) are passed and they are not
added to the interwikimap.

<!--
This PR fixes the Issue
-->

- [x] 🍕 Feature
- [x] ✅ Test

Fixes #79

- [x] 👍 yes

- [ ] 📜 README.md
- [ ] 📖 own file under the docs folder
- [x] 🙅 no documentation needed

<!--
Write about changing templates or modules on-wiki in order to accomodate
changes in this PR
-->