Returned Elokuun kuvitustalkoot config back to "fi-ek" from "fi-ek202… (#88)
…3" in jobs.yaml

## Description

There was a custom Elokuun kuvitustalkoot config in 2023, probably
because of incorrect competition settings, and it was worked around by
using a different competition ID. However, since it only worked in 2023,
I have now reverted the jobs.yaml back to the multiyear config.

NOTE: This is also updated now directly to production as the competition
is already started.